Rename the CurvesStatsWidget to StatsWidget only
