+++ silx/gui/plot/StatsWidget.py
