+++ silx/gui/plot/test/__init__.py
@@ -32,7 +32,7 @@ import unittest
 from .._utils import test
 from . import testColorBar
 from . import testCurvesROIWidget
-from . import testCurvesStatsWidget
+from . import testStatsWidget
 from . import testAlphaSlider
 from . import testInteraction
 from . import testLegendSelector
@@ -59,7 +59,7 @@ def suite():
         [test.suite(),
          testColorBar.suite(),
          testCurvesROIWidget.suite(),
-         testCurvesStatsWidget.suite(),
+         testStatsWidget.suite(),
          testAlphaSlider.suite(),
          testInteraction.suite(),
          testLegendSelector.suite(),
